requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
docstring-parser==0.14.1
fastapi==0.78.0
flake8==4.0.1
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
Jinja2==3.1.2
jsonargparse==4.9.0
lightning-flash==0.7.5
Markdown==3.3.7
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.1
pyDeprecate==0.3.2
pyflakes==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.6.4
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.11.0
torchmetrics==0.9.1
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
